Preserve example documents and tests for previous version (2023-03) of CSB schemas
