+++ docs/NOAA/noaa_b12_v3_0_0_xyz_suggested-2023-03.json
@@ -1,0 +1,141 @@
+{
+  "crs": {
+    "horizontal": {
+      "type": "EPSG",
+      "value": 8252
+    },
+    "vertical": "Transducer"
+  },
+  "providerContactPoint": {
+    "orgName": "Example Cruises Inc",
+    "email": "support@example.com",
+    "logger": "Rose Point ECS",
+    "loggerVersion": "1.0"
+  },
+  "convention": "XYZ CSB 3.0",
+  "dataLicense": "CC0 1.0",
+  "platform": {
+    "uniqueID": "EXAMPLE-f8c469f8-df38-11e5-b86d-9a79f06e9478",
+    "type": "Ship",
+    "name": "USCGC Healy",
+    "length": 128,
+    "IDType": "MMSI",
+    "IDNumber": 908338000,
+    "sensors": [
+      {
+        "type": "Sounder",
+        "make": "Kongsberg Maritime",
+        "model": "EM122",
+        "position": [
+          -18.404,
+          -1.909,
+          8.919
+        ],
+        "draft": 8.321,
+        "frequency": 12000,
+        "transducer": "EM122-CentreBeam"
+      },
+      {
+        "type": "GNSS",
+        "make": "Kongsberg Maritime",
+        "model": "Seapath 330+",
+        "position": [
+          0,
+          0,
+          0
+        ]
+      },
+      {
+        "type": "MotionSensor",
+        "make": "Kongsberg Maritime",
+        "model": "Seapath 330+",
+        "position": [
+          0,
+          0,
+          0
+        ]
+      }
+    ],
+    "correctors": {
+      "positionReferencePoint": "GNSS",
+      "soundSpeedDocumented": true,
+      "positionOffsetsDocumented": true,
+      "dataProcessed": true,
+      "motionOffsetsApplied": true,
+      "draftApplied": true
+    }
+  },
+  "algorithms": [
+    {
+      "name": "deduplicate",
+      "params": ""
+    }
+  ],
+  "lineage": [
+    {
+      "type": "CRSTransformation",
+      "timestamp": "2021-11-22T16:10:09.346821",
+      "detail": {
+        "origin": "EPSG:4326",
+        "destination": "EPSG:8252",
+        "method": "GeoTrans"
+      }
+    },
+    {
+      "type": "TimeInterpolation",
+      "timestamp": "2021-11-22T16:10:09.346821",
+      "detail": {
+        "source": "SystemTime | ZDA | RMC | ...",
+        "method": "Midway | ElapsedTime | DataSource | NTP | PTP | ...",
+        "comment": "Millisecond elapsed time on first character of strings used for interpolation"
+      }
+    },
+    {
+      "type": "algorithm",
+      "timestamp": "2021-11-22T16:10:09.346821",
+      "detail": {
+        "name": "deduplicate",
+        "parameters": "",
+        "comment": "Selected 11281 non-duplicate depths from 35744 in input."
+      }
+    },
+    {
+      "type": "GNSSProcessing",
+      "timestamp": "2021-11-22T16:10:09.346821",
+      "detail": {
+        "algorithm": "RTKLib | CSRS-PPP | ...",
+        "version": "1.2.0",
+        "comment": "Base stations from New Hampshire CORS at NHUN https://www.unh.edu/facilities/gis-cors"
+      }
+    },
+    {
+      "type": "SoundSpeedCorrection",
+      "timestamp": "2021-11-22T16:10:09.346821",
+      "detail": {
+        "source": "Model | Profile | GeometricMean | Fixed | CarterTable | MatthewTable | WOA13 | ...",
+        "algorithm": "CIDCO-Ocean | Raytraced | ...",
+        "version": "1.1.0",
+        "comment": "Solutions from 25 profiles, 15 used, 10 duplicate."
+      }
+    },
+    {
+      "type": "VerticalReduction",
+      "timestamp": "2021-11-22T16:10:09.346821",
+      "detail": {
+        "reference": "ChartDatum | Ellipsoid",
+        "method": "ObservedWL | PredictedWL | ModelledWL | EllipsoidReduction | ...",
+        "model": "CANNORTH2016v1HyVSEP_NAD83v6_CD | StationReference | VDatum | VORF | ...",
+        "comment": "Water levels from NOAA gauge https://tidesandcurrents.noaa.gov/stationhome.html?id=8419870"
+      }
+    },
+    {
+      "type": "CRS",
+      "timestamp": "2021-11-22T16:10:09.346821",
+      "detail": {
+        "type": "EPSG",
+        "value": 8252,
+        "comment": "Transformed with GeoTrans v 3.23"
+      }
+    }
+  ]
+}
